- updated the UnitsNet.Benchmark project (updated nugets & added frameworks)
- added BenchmarkCategories to the existing benchmarks
- added a few scripts for local testing (individual, multi-platform + rplots)
- added an automatic benchmark workflow on pushing to [master] (some folders), running for ["netcoreapp50", "netcoreapp21", "net472"]
- added workflows triggered on workflow_dispatch for running a benchmark (with options for comparing against a baseline)

Closes #920 

First a few ground rules

1. As of this writing, manually triggered workflows (a.k.a. `workflow_dispatch`), [only work on the master branch](https://github.community/t/how-to-trigger-repository-dispatch-event-for-non-default-branch/14470). 
2. Composite actions have some nasty [limitations](https://github.com/actions/runner/issues/862)
3. The Continuous Benchmark [needs to be in a context](https://github.com/rhysd/github-action-benchmark/issues/24) with a commit in order to work 
4.  It publishes [commits](https://github.com/lipchev/UnitsNet-Benchmarks/commits/gh-pages/benchmarks/netcoreapp50) (with optional comments and alerts) targeting a given folder on a separate branch (typically gh-pages/benchmark/)- [here](https://lipchev.github.io/UnitsNet-Benchmarks/benchmarks/netcoreapp50/) is an example (all but the last benchmarks were run using a very-low iteration count- a sort of DryRun, so the wild fluctuations are expected). Here is [another](https://lipchev.github.io/UnitsNet-Benchmarks/benchmarks/net472/) (that's consistently slower).
5. It [does not support](https://github.com/rhysd/github-action-benchmark/issues/18) multi-line charts (out-of-the-box)
6. I read somewhere that the average performance of the virtual machines vary within about 10-15%
7. We can run the benchmarks for [any framework](https://benchmarkdotnet.org/articles/configs/toolchains.html#multiple-frameworks-support) from 472 onward (windows required)
8. We can easily configure all aspects of a benchmark using the [command line arguments](https://benchmarkdotnet.org/articles/guides/console-args.html) (there is also a `dotnet benchmark tool` but that seems to be doing the exact same things- although I suspect it's also going to get some 'compare results' functionality at some point which would likely remove the need for the next step)
9. There are a few result-comparing extensions out there- but I most liked the one from [dotnet/performance](https://github.com/dotnet/performance/tree/main/src/tools/ResultsComparer)
10. Just a few days ago, I knew almost none of that.

The workflows are well documented, I think you would easily figure out what's going on. Here are the highlights:

1. The automatic workflow runs the benchmarks (multiple frameworks), for each one exporting the result to the [gh-pages/benchmarks](https://github.com/lipchev/UnitsNet-Benchmarks/tree/gh-pages/benchmarks)- one framework one sub-folder
2. It is only supposed to run on the master branch & only on certain occasions (see [here](https://github.com/lipchev/UnitsNet/blob/dd071c6255424049401ad1d2ea6c7c85383eb38c/.github/workflows/continious-benchmarking.yml#L2))
2. For [each folder](https://github.com/lipchev/UnitsNet-Benchmarks/tree/gh-pages/benchmarks/net472) (framework) we have the chart data: `data.js & index.html` (here we append results) and the [results](https://github.com/lipchev/UnitsNet-Benchmarks/tree/gh-pages/benchmarks/net472/results) folder containing the latest benchmark results, as exported by BenchmarkDotNet (we override these each time the script is executed)
3. We can then reference / compare results across branches & repositories (a fork doesn't have to have gh-pages in order to compare against upstream/gh-pages/benchmark/xxx/)
4. This can be done using the second action: which can be used to generate a report (the typical results folder) which can then be compared against a baseline (this step only runs if the [baseline parameter](https://github.com/lipchev/UnitsNet/blob/dd071c6255424049401ad1d2ea6c7c85383eb38c/.github/workflows/run-benchmarks.yml#L29) points to a valid "*-report-full.json")

PS Some of the links point to lipchev/UnitsNet-Benchmarks which is the repository I used for testing. Unfortunately something got messed up during the rebase so I made the PR in a new branch (of my main fork) - this (as expected) did not trigger a benchmark-run as I pushed on a non master branch. As to why the manual flow isn't showing up in the actions menu- see 1). 

PS2 You should still have access to the test repository (lipchev/UnitsNet-Benchmarks) if you want to test out the [workflows](https://github.com/lipchev/UnitsNet-Benchmarks/actions)- feel free to commit on it if you want, I plan to delete it after this is merged.