@angularsen 2021-04-30
I mean, wow, great job and great outline! I love it ❤️ 
I say, pull the trigger! 

Letting you merge it, in case there is anything else to do before it merges.

The only thing I'm curious about, is how do we best use this tool to monitor regressions over time? 
If I understand correctly, this only runs on `master` so we can't easily use it when reviewin PRs. Is there any way to alert us to give a heads up if it regresses significantly, say 50% slower on some tests?

Also, are there any limits on the free plan for actions we need to know about?