@lipchev 2021-04-30
Ok, here are the initial results:
https://angularsen.github.io/UnitsNet/benchmarks/net472/
https://angularsen.github.io/UnitsNet/benchmarks/netcoreapp21/
https://angularsen.github.io/UnitsNet/benchmarks/netcoreapp50/

And the detailed results:
https://github.com/angularsen/UnitsNet/tree/gh-pages/benchmarks/net472/results
https://github.com/angularsen/UnitsNet/tree/gh-pages/benchmarks/netcoreapp21/results
https://github.com/angularsen/UnitsNet/tree/gh-pages/benchmarks/netcoreapp50/results