@lipchev 2021-05-01
Ok, here is an example run from my newly created fork that returns a new list for each unknown abbreviation:
![image](https://user-images.githubusercontent.com/5253184/116766248-2faebe80-aa32-11eb-9657-4dd5a38b62a1.png)

[Here](https://github.com/lipchev/UnitsNet/actions/runs/801146843) is the action log (net50)

And here are the results:

Runtime=.NET Core 5.0  Toolchain=netcoreapp50  

```
|          Method |     Mean |    Error |   StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|---------------- |---------:|---------:|---------:|-------:|-------:|------:|----------:|
|           Parse | 49.42 μs | 0.980 μs | 1.048 μs | 1.7700 | 0.0610 |     - |  32.56 KB |
|   TryParseValid | 49.98 μs | 0.736 μs | 0.653 μs | 1.7700 | 0.0610 |     - |  32.54 KB |
| TryParseInvalid | 54.05 μs | 1.061 μs | 1.090 μs | 1.7090 | 0.0610 |     - |  32.16 KB |

..and the results of the comparison to the baseline
* (not comparing the allocations, but if you open the base-line from the link above you should be able to spot the difference)

summary:
worse: 3, geomean: 1.072
total diff: 3

| Slower                                                | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| UnitsNet.Benchmark.UnitsNetBenchmarks.TryParseInvalid |      1.10 |         48927.15 |         53800.22 |         |
| UnitsNet.Benchmark.UnitsNetBenchmarks.TryParseValid   |      1.08 |         46153.61 |         49879.57 |         |
| UnitsNet.Benchmark.UnitsNetBenchmarks.Parse           |      1.04 |         47511.59 |         49251.15 |         |

No Faster results for the provided threshold = 0.001% and noise filter = 0.3ns.