@lipchev 2021-05-01
Ok I only needed to fix [one little thing](https://github.com/angularsen/UnitsNet/commit/3bca803d0beb9c6e5da35b250d25c3aa09ddc807). This in turn triggered a re-run, so now we should get two points on the chart.

Overall the benchmarks appear to be quite noisy- some of the operations seem to be too short. For those cases microsoft is doing a [bunch of operations in a loop](https://github.com/dotnet/performance/blob/d7dac8a7ca12a28d099192f8a901cf8e30361384/src/benchmarks/micro/runtime/Math/Functions/Double/MinDouble.cs). I'm thinking of stealing this test, and putting it side by side with ours (sort of like a baseline for the IComparable interface, for which there is currently no benchmark).