@angularsen 2021-05-07
I guess it might be tricky to get stable numbers from a shared pool of worker VMs.

This one observed 10-20% fluctuation.
https://github.com/rhysd/github-action-benchmark#stability-of-virtual-environment