@lipchev 2021-05-07
Yeah, pretty useless for any between-commit comparisons (might look ok in the long run, though I think I'm going to disable the auto-comment feature).. 

Also- I hadn't foreseen it triggering on release, but yeah- I guess the project file is inside the UnitsNet folder :)

Anyways, we got lucky this time- [Intel Xeon CPU E5-2673 v3 2.40GHz appears to have almost the same performance as Intel Xeon Platinum 8171M CPU 2.60GHz](benchmarks/netcoreapp50/results/UnitsNet.Benchmark.UnitsNetBenchmarks-report-github.md)

PS I'm playing with a little _hack_ for re-scaling the results (only for the charts)- that would probably smooth out the lines a little. 

PS1 There is a bug in the QuantityFrom benchmark (I've fixed it locally).

PS2 There are at least two performance improvements I'm going to PR for soon (so that this wouldn't have been a complete waste of time..).